+++ src/tasks/LoopCharacteristics.js
@@ -63,6 +63,10 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
     }
 
     function executeSequential() {
+      const {cardinality, getContent: getStartContent} = getCharacteristics();
+      if (cardinality === 0) return complete();
+      if (!cardinality && !startCondition && !completionCondition) return activity.emitFatal(new ActivityError(`<${id}> cardinality, collection, or condition is required in sequential loops`, executeMessage), getStartContent());
+
       let startIndex = 0;
 
       if (isRedelivered && executeRoutingKey === 'execute.iteration.next') {
@@ -115,11 +119,18 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
 
         debug(`<${parentExecutionId} (${id})> sequential loop completed`);
 
+        return complete(content);
+      }
+
+      function complete(content) {
         stop();
+
+        const {getContent, output} = getCharacteristics();
+
         return broker.publish('execution', 'execute.completed', {
-          ...message.content,
-          ...getCharacteristics().getContent(),
-          output: loopOutput,
+          ...content,
+          ...getContent(),
+          output,
         });
       }
     }
@@ -277,11 +288,12 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
     }
 
     function getCardinality(collection) {
+      const collectionLen = Array.isArray(collection) ? collection.length : undefined;
       if (!loopCardinality) {
-        return Array.isArray(collection) ? collection.length : undefined;
+        return collectionLen;
       }
       const value = environment.resolveExpression(loopCardinality, executeMessage);
-      if (value === undefined) return;
+      if (value === undefined) return collectionLen;
       return Number(value);
     }